Change github release plugin and update build scripts
