+++ tachikoma-backend-api-proto/tachikoma-backend-api-jvm/build.gradle.kts
@@ -1,7 +1,7 @@
 import com.google.protobuf.gradle.protobuf
 
 plugins {
-    id("tachikoma.grpc")
+    `tachikoma-grpc`
 }
 
 evaluationDependsOn(":tachikoma-backend-api-proto")
